Use Int as of passing in URL

+++ Keyhac/ExtensionApiLayer/KeyhacCore_Chooser.swift
@@ -55,7 +55,7 @@ public class Chooser {
         self.destroy()
     }
     
-    public func open( x: Double, y: Double, width: Double, height: Double ) {
+    public func open( x: Int, y: Int, width: Int, height: Int ) {
         if let url = URL(string: "keyhac://chooser/\(name)?x=\(x)&y=\(y)&width=\(width)&height=\(height)") {
             NSWorkspace.shared.open(url)
         }
